Create README.md
This project will be focused on cleaning and analyzing exit surveys from employees of the Department of Education, Training and Employment and the Technical and Further Education (TAFE) body of the Queensland government in Australia. The TAFE exit survey can be found here and the survey for the DETE can be found here.

As a data analyst, our stakeholders want us to combine the results for both surveys to answer the following questions:

Are employees who only worked for the institutes for a short period of time resigning due to some kind of dissatisfaction? What about employees who have been there longer?

Are younger employees resigning due to some kind of dissatisfaction? What about older employees?